Movement Updates (#517)
* Movement Updates

* Fix combat stance bug and added acecommands tip to console at startup

* Update changelog.md

* Fixed palette issue with starter head gear

* The Code Alignment Gods have blessed the enumerations, and it is good.

+++ Source/ACE/Entity/Door.cs
@@ -52,8 +52,8 @@ namespace ACE.Entity
             if (IsLocked)
                 DefaultLocked = true;
 
-            movementOpen.ForwardCommand = (ushort)MotionCommand.On;
-            movementClosed.ForwardCommand = (ushort)MotionCommand.Off;
+            movementOpen.ForwardCommand = (uint)MotionCommand.On;
+            movementClosed.ForwardCommand = (uint)MotionCommand.Off;
 
             if (UseRadius < 2)
                 UseRadius = 2;
